% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortuuid.R
\name{bitcoin58_to_uuid}
\alias{bitcoin58_to_uuid}
\title{Convert base58 bitcoin encoded character vector to uuid}
\usage{
bitcoin58_to_uuid(input)
}
\arguments{
\item{input}{character vector of base58 strings}
}
\value{
character vector of uuids
}
\description{
Convert base58 bitcoin encoded character vector to uuid
}
