% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortuuid.R
\name{generate_uuid}
\alias{generate_uuid}
\title{Generate a random RFC4122 v4-compliant UUID}
\usage{
generate_uuid(n = 1)
}
\arguments{
\item{n}{number of ids to generate}
}
\value{
character vector of uuids
}
\description{
Generate a random RFC4122 v4-compliant UUID
}
\examples{
generate_uuid(n = 5)

}
