% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.uuid}
\alias{is.uuid}
\title{check if object is of class uuid}
\usage{
is.uuid(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
logical indicating if the input is a valid UUID
}
\description{
check if object is of class uuid
}
