% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate.uuid}
\alias{validate.uuid}
\title{validate if a string is a uuid}
\usage{
validate.uuid(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
logical indicating if the input is a valid UUID
}
\description{
validate if a string is a uuid
}
