% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortuuid.R
\name{base58_to_uuid}
\alias{base58_to_uuid}
\title{Convert base58 to uuid}
\usage{
base58_to_uuid(input, alphabet)
}
\arguments{
\item{input}{base58}

\item{alphabet}{character representing alphabet}
}
\description{
Convert base58 to uuid
}
