% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.base58}
\alias{is.base58}
\title{validate if string is base58}
\usage{
is.base58(x, alphabet)
}
\arguments{
\item{x}{a character vector to check}

\item{alphabet}{character vector representing alphabet}
}
\description{
validate if string is base58
}
