% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.base58}
\alias{is.base58}
\title{validate if character vector is base58 encoded}
\usage{
is.base58(x, alphabet)
}
\arguments{
\item{x}{A character vector}

\item{alphabet}{character vector representing an alphabet}
}
\description{
validate if character vector is base58 encoded
}
