% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortuuid.R
\name{uuid_to_base58}
\alias{uuid_to_base58}
\title{Convert uuid to base58}
\usage{
uuid_to_base58(input, alphabet)
}
\arguments{
\item{input}{uuid}

\item{alphabet}{character vector representing alphabet}
}
\description{
Convert uuid to base58
}
